+++ appcontainer/ViewportSizeModel.js
@@ -5,8 +5,10 @@
  * Copyright © 2021 Extremely Heavy Industries Inc.
  */
 import {HoistModel} from '@xh/hoist/core';
-import {action, computed, makeObservable, observable} from '@xh/hoist/mobx';
-import {isFinite} from 'lodash';
+import {action, makeObservable, observable} from '@xh/hoist/mobx';
+import {debounced} from '@xh/hoist/utils/js';
+import {isFinite, isString} from 'lodash';
+import {runInAction} from 'mobx';
 
 /**
  * Track observable width / height of the browser viewport, and provide observable
@@ -19,20 +21,10 @@ export class ViewportSizeModel extends HoistModel {
     /** @member {Object} - contains `width` and `height` in pixels */
     @observable.ref size;
 
-    /** @returns {boolean} */
-    @computed
-    get isPortrait() {
-        const {size} = this; // Force triggering observation of size.
+    /** @member {boolean} */
+    @observable isPortrait;
 
-        // Check Modern API and legacy API (for safari, BB Access)
-        let orientation = this.getOrientation() ?? this.getLegacyOrientation();
-        if (orientation !== null) return orientation === 0 || orientation === 180;
-
-        // Default to aspect ratio
-        return size.width < size.height;
-    }
-
-    /** @returns {boolean} */
+    /** @returns {boolean} - observable inverse of isPortrait. */
     get isLandscape() {
         return !this.isPortrait;
     }
@@ -42,27 +34,41 @@ export class ViewportSizeModel extends HoistModel {
         makeObservable(this);
         window.addEventListener('resize', () => this.setViewportSize());
         this.setViewportSize();
+
+        this.addReaction({
+            track: () => this.size,
+            run: () => this.updateOrientation(),
+            fireImmediately: true,
+            // Further debounce (on top of debounced setViewportSize) to allow for even greater
+            // lag observed on mobile device responses to DOM orientation APIs.
+            debounce: 100
+        });
     }
 
     //---------------------
     // Implementation
     //---------------------
-    @action
+    @debounced(100)  // Debounced to account for slow inner window resize on mobile.
     setViewportSize() {
-        this.size = {
+        runInAction(() => this.size = {
             width: window.innerWidth,
             height: window.innerHeight
-        };
+        });
     }
 
-    getOrientation() {
-        const {orientation} = window.screen;
-        return orientation ? orientation.angle : null;
-    }
-
-    getLegacyOrientation() {
-        const {orientation} = window;
-        return isFinite(orientation) ? orientation : null;
+    @action
+    updateOrientation() {
+        if (isString(window.screen?.orientation?.type)) {
+            // First check modern API.
+            this.isPortrait = window.screen.orientation.type.startsWith('portrait');
+        } else if (isFinite(window.orientation)) {
+            // ...fallback to older API for Safari/BB browser.
+            this.isPortrait = window.orientation === 0 || window.orientation === 180;
+        } else {
+            // ...final fallback to aspect ratio.
+            const {width, height} = this.size;
+            this.isPortrait = width < height;
+        }
     }
 
 }